requirements.txt
pandas
python-dotenv
sentence-transformers
faiss-cpu
torch
groq
openai
google-generativeai
tqdm
nltk

langchain-text-splitters
numpy
transformers
accelerate
bitsandbytes